requirements.txt
django 
channels
channels-redis
grpcio
grpcio-tools 
djangorestframework 
django-ratelimit
redit